Move exceptions to the `org.stellar.anchor.api.exception` package.
