+++ api-schema/src/main/java/org/stellar/anchor/api/exception/Sep31CustomerInfoNeededException.java
@@ -1,0 +1,13 @@
+package org.stellar.anchor.api.exception;
+
+public class Sep31CustomerInfoNeededException extends AnchorException {
+  private final String type;
+
+  public Sep31CustomerInfoNeededException(String type) {
+    this.type = type;
+  }
+
+  public String getType() {
+    return type;
+  }
+}
