+++ platform/src/main/java/org/stellar/anchor/platform/controller/Sep31Controller.java
@@ -9,13 +9,13 @@ import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
 import org.springframework.web.bind.annotation.*;
 import org.stellar.anchor.api.exception.AnchorException;
+import org.stellar.anchor.api.exception.Sep31CustomerInfoNeededException;
+import org.stellar.anchor.api.exception.Sep31MissingFieldException;
 import org.stellar.anchor.api.sep.AssetInfo.Sep31TxnFieldSpecs;
 import org.stellar.anchor.api.sep.sep31.*;
 import org.stellar.anchor.auth.JwtToken;
 import org.stellar.anchor.platform.condition.ConditionalOnAllSepsEnabled;
 import org.stellar.anchor.sep31.Sep31Service;
-import org.stellar.anchor.sep31.Sep31Service.Sep31CustomerInfoNeededException;
-import org.stellar.anchor.sep31.Sep31Service.Sep31MissingFieldException;
 
 @RestController
 @CrossOrigin(origins = "*")
